3 performance fixes in YOLO post-processing (#3915)
- in detect_postprocess, scores is a writeable slice. When exporting to
.pt2, this generates a ScatterND layer that hurts performance on all
platforms, but mostly on TFLite
- in get_most_likely_score, issue #8595 appears to be fixed. No more
need to Split/Concat around the ReduceMax layer
- in get_most_likely_score, remove the cast to int, as all but one
caller already inserts a cast. Fix the one caller that did not

<!--
  PR title must include a JIRA ticket, e.g.: TETRAAI-123 Add feature X
  CI will fail if the title does not contain a valid TETRAAI-### key.
-->

## Summary

<!-- Describe your changes here -->

## Test Plan

<!-- How did you verify this change? -->

---------

+++ src/qai_hub_models/models/_shared/yolo/model.py
@@ -124,7 +124,7 @@ def yolo_segment_postprocess(
     # Get class ID of most likely score.
     scores, class_idx = get_most_likely_score(scores)
 
-    return boxes, scores, masks, class_idx
+    return boxes, scores, masks, class_idx.to(torch.int32)
 
 
 def yolo_obb_postprocess(
